requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# project specific requirements
awpy==1.1.6
inflection==0.5.1
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.2
python-dotenv==0.20.0
scikit_learn==1.1.1
seaborn==0.11.2
setuptools==62.1.0
tqdm==4.64.0